vdev GraphCommons

Core module for the ExGraphsBook umbrella project.

This module serves as the entry point for the documentation as well as providing the basic structs %Graph{} and %Query{} in GraphCommons.Graph and GraphCommons.Query, respectively.

Types

base_type()

Specs

base_type() :: :dgraph | :native | :property | :rdf | :tinker

Types for graph storage.

file_test()

Specs

file_test() :: :dir? | :regular? | :exists?

Type for testing file types.

graph_type()

Specs

graph_type() :: base_type()

query_type()

Specs

query_type() :: base_type()

Functions

exports_dir()

Specs

exports_dir() :: String.t()

Returns the common exports area under the :graph_commons app.

hello()

Specs

hello() :: :ok

Hello world for the ExGraphsBook umbrella app.

Examples

iex> GraphCommons.hello()
This is ExGraphsBook - an Elixir umbrella app:

[:graph_browser, :graph_commons, :graph_connect, :graph_share, :graph_transform,
 :native_graph, :property_graph, :rdf_graph, :tinker_graph]
:ok

imports_dir()

Specs

imports_dir() :: String.t()

Returns the common imports area under the :graph_commons app.

scripts_dir()

Specs

scripts_dir() :: String.t()

Returns the common scripts area under the :graph_commons app.

storage_dir()

Specs

storage_dir() :: String.t()

Returns the common storage area under the :graph_commons app.