vdev GraphCommons.Graph

Module providing a struct for graphs.

The %GraphCommons.Graph{} struct is used to bundle related data access fields for graph serializations together.

One of the fields :type tracks graph_type:

:native – native graph

:property – property graph

:rdf – RDF graph

:tinker – TinkerPop graph

The other fields are straightforward data access fields:

:data – graph_data

:file – name of graph_file (within the graphs_dir)

:path – absolute path of graph_file

:uri – absolute URI of graph_file

Types

file_test()

Specs

file_test() :: GraphCommons.file_test()

graph_data()

Specs

graph_data() :: String.t()

Type for graph data string in graph storage.

graph_file()

Specs

graph_file() :: String.t()

Type for file name of graph storage.

graph_path()

Specs

graph_path() :: String.t()

Type for file path of graph storage.

graph_type()

Specs

graph_type() :: GraphCommons.graph_type()

Type for type of graph in graph storage.

graph_uri()

Specs

graph_uri() :: String.t()

Type for (file:) URI of graph storage.

t()

Specs

t() :: %GraphCommons.Graph{
  data: graph_data(),
  file: graph_file(),
  path: graph_path(),
  type: graph_type(),
  uri: graph_uri()
}

Type for %GraphCommons.Graph{} struct.

Functions

is_file_type(file_type)
(macro)

Guard to test for valid file_type.

is_graph_type(graph_type)
(macro)

Guard to test for valid graph_type.

list_graphs(graph_type, file_test \\ :exists?)

Specs

list_graphs(graph_type(), file_test()) :: [String.t()]

Returns a (sorted) listing of the graphs directory, optionally filtered by file_test.

list_graphs_dir(graph_file, graph_type, file_test \\ :exists?)

Specs

list_graphs_dir(graph_file(), graph_type(), file_test()) :: [String.t()]

Returns a (sorted) listing of a named file in the graphs directory, optionally filtered by file_test.

new(graph_data, graph_file, graph_type)

Specs

new(graph_data(), graph_file(), graph_type()) :: t()

Creates a %GraphCommons.Graph{} struct.

Examples

iex> "CREATE (a)-[:EX]->(b)" |>
...> GraphCommons.Graph.new("test.cypher", :property)
#GraphCommons.Graph<type: property, file: "test.cypher", data: "CREATE (a)-[:EX]...">

iex> "CREATE (TheMatrix:Movie {title:'The Matrix', released:1999})\n" |>
...> GraphCommons.Graph.new("test.cypher", :property)
#GraphCommons.Graph<type: property, file: "test.cypher", data: "CREATE (TheMatri...">

read_graph(graph_file, graph_type)

Specs

read_graph(graph_file(), graph_type()) :: t()

Returns a new graph with graph_data after reading storage for graph_file and graph_type.

write_graph(graph_data, graph_file, graph_type)

Specs

write_graph(graph_data(), graph_file(), graph_type()) :: t()

Returns a new graph with graph_data after writing storage for graph_data, graph_file and graph_type.