vdev GraphCommons.Query

Module providing a struct for queries.

The %GraphCommons.Query{} struct is used to bundle related data access fields for query serializations together.

One of the fields :type tracks query_type:

:native – native query

:property – property query

:rdf – RDF query

:tinker – TinkerPop query

The other fields are straightforward data access fields:

:data – query_data

:file – name of query_file (within the queries_dir)

:path – absolute path of query_file

:uri – absolute URI of query_file

Types

file_test()

Specs

file_test() :: GraphCommons.file_test()

Type for testing file types.

query_data()

Specs

query_data() :: String.t()

Type for query data string in query storage.

query_file()

Specs

query_file() :: String.t()

Type for file name of query storage.

query_path()

Specs

query_path() :: String.t()

Type for file path of query storage.

query_type()

Specs

query_type() :: GraphCommons.query_type()

Type for type of query in query storage.

query_uri()

Specs

query_uri() :: String.t()

Type for (file:) URI of query storage.

t()

Specs

t() :: %GraphCommons.Query{
  data: query_data(),
  file: query_file(),
  path: query_path(),
  type: query_type(),
  uri: query_uri()
}

Type for %GraphCommons.Query{} struct.

Functions

is_type(query_type)
(macro)

list_queries(query_type, file_test \\ :exists?)

Specs

list_queries(query_type(), file_test()) :: [String.t()]

Returns a listing of the queries directory, optionally filtered by file_test.

list_queries_dir(query_file, query_type, file_test \\ :exists?)

Specs

list_queries_dir(query_file(), query_type(), file_test()) :: [String.t()]

Returns a listing of a named file in the queries directory, optionally filtered by file_test.

new(query_data, query_file, query_type)

Specs

new(query_data(), query_file(), query_type()) :: t()

Creates a %GraphCommons.Query{} struct.

Examples

iex> query = RDFGraph.read_query
#GraphCommons.Query<type: rdf, file: "default.rq", data: "construct\n{ ?s ?p ?o }\...">

iex> query.data |> GraphCommons.Query.new("test.rq", :rdf)
#GraphCommons.Query<type: rdf, file: "test.rq", data: "construct\n{ ?s ?p ?o }\...">

read_query(query_file, query_type)

Specs

read_query(query_file(), query_type()) :: t()

Returns a new query with query_data after reading storage for query_file and query_type.

write_query(query_data, query_file, query_type)

Specs

write_query(query_data(), query_file(), query_type()) :: t()

Returns a new query with query_data after writing storage for query_data, query_file and query_type.