vdev GraphCommons.Service behaviour

Module defining graph service behaviour callbacks.

This behaviour defines a CRUD interface to a graph service with the following callbacks:

graph_create/1

graph_read/0

graph_update/1

graph_delete/0

Added to that basic set is an optional callback:

graph_info/0

And lastly we also define a couple callbacks for querying the graph:

query_graph/1

query_graph!/1

For interactive use the following function helpers are provided:

GraphCommons.Utils.cypher!/1

GraphCommons.Utils.gremlin!/1

GraphCommons.Utils.libgraph!/1

GraphCommons.Utils.sparql!/1

These latter will typically be imported so that they can be invoked with no namespacing required, e.g cypher/1, gremlin!/1, libgraph!/1, sparql!/1.

Callbacks

graph_create(arg1)

Specs

graph_create(GraphCommons.Graph.t()) :: any()

Creates graph in the service graph store.

graph_delete()

Specs

graph_delete() :: any()

Deletes graph in the service graph store.

graph_info()
(optional)

Specs

graph_info() :: any()

Reports on graph in the service graph store.

graph_read()

Specs

graph_read() :: any()

Reads graph from the service graph store.

graph_update(arg1)

Specs

graph_update(GraphCommons.Graph.t()) :: any()

Writes graph to the service graph store.

query_graph(arg1)

Specs

query_graph(GraphCommons.Query.t()) :: any()

Queries graph in the service graph store.

query_graph(arg1, map)
(optional)

Specs

query_graph(GraphCommons.Query.t(), map()) :: any()

Queries graph in the service graph store.

query_graph!(arg1)

Specs

query_graph!(GraphCommons.Query.t()) :: any()

Queries graph in the service graph store with direct result.

query_graph!(arg1, map)
(optional)

Specs

query_graph!(GraphCommons.Query.t(), map()) :: any()

Queries graph in the service graph store with direct result.