vdev GraphCommons.Utils

Module providing helper functions.

Functions

close_socket(socket)

cypher(query_string)

Specs

cypher(String.t()) :: any()

cypher!(query_string)

Specs

cypher!({String.t(), map()}) :: any()

cypher!(String.t()) :: any()

Queries a PropertyGraph graph store with a query_string string arg.

cypher!(query_string, query_params)

Specs

cypher!(String.t(), map()) :: any()

Queries a PropertyGraph graph store with a query_string string arg and a query_params map arg.

dgraph!(query_string)

Specs

dgraph!(String.t()) :: any()

eval(string)

Specs

eval(String.t()) :: any()

Returns result of evaluating a string.

eval_script(script)

Specs

eval_script(String.t()) :: any()

Returns result of evaluating a script bundled with the :graph_commons app.

graph_modules()

Specs

graph_modules() :: keyword()

Generates keyword list of app, module pairs from graph modules used in this umbrella app.

graph_service(graph_module)
(macro)

Specs

graph_service(module()) :: module()

graph_service(module()) :: GraphCommons.Service.GraphInfo.t()

Macro to set graph service with graph_module.

gremlin!(query_string)

Specs

gremlin!(String.t()) :: any()

Queries a TinkerGraph graph store with a query_string arg.

gs(graph_module)
(macro)

help()

Specs

help() :: :ok

Prints out ...

Examples

iex> help
[:graph_commons]
:ok

help(arg)

Specs

help(atom()) :: :ok

Prints out ...

is_module(graph_module)
(macro)

libgraph!(query_string)

Specs

libgraph!(String.t()) :: any()

Queries a NativeGraph graph store with a query_string arg.

list_apps()

Specs

list_apps() :: :ok

Prints out apps used in this umbrella app.

Examples

iex> list_apps
[:graph_browser, :graph_commons, :graph_connect, :graph_transform,
 :native_graph, :property_graph, :rdf_graph]
:ok

list_functions(module)

Specs

list_functions(module()) :: :ok

Prints out functions used in module.

Examples

iex> list_functions(GraphCommons)
[GraphCommons, GraphCommons.Graph, GraphCommons.Query, GraphCommons.Utils,
 Inspect.GraphCommons.Graph, Inspect.GraphCommons.Query]
:ok

list_graph_modules()

Specs

list_graph_modules() :: [module(), ...]

Prints out toplevel graph modules used in this umbrella app.

list_macros(module)

Specs

list_macros(module()) :: :ok

Prints out macros used in module.

Examples

iex> list_macros(GraphCommons.Utils)
[graph_service: 1, gs: 1, is_module: 1]
:ok

list_modules(app)

Specs

list_modules(atom()) :: :ok

Prints out modules used in app.

Examples

iex> list_mods(:graph_commons)
[help: 0, list_apps: 0, list_functions: 1, list_modules: 1]
:ok

open_socket(address, port)

repeat(times, block)
(macro)

sparql(query_string)

Specs

sparql(String.t()) :: any()

Queries a RDFGraph graph store with a query_string arg.

sparql!(query_string)

Specs

sparql!(String.t()) :: any()

Queries a RDFGraph graph store with a query_string arg.

sparql!(query_string, query_params)

Specs

sparql!(String.t(), map()) :: any()

Queries a RDFGraph graph store with a query_string string arg and a query_params map arg.