vdev PropertyGraph

Module for reading and writing a library of property graphs and Cypher queries.

The read_graph/1 and write_graph/2 functions allow for reading and writing RDF graphs to the project data repository. (Default file names are provided with the read_graph/0 and write_graph/1 forms.) The list_graphs/0 function lists graph file names.

The read_query/1 and write_query/2 functions allow for reading and writing SPARQL queries to the project data repository. (Default file names are provided with the read_query/0 and write_query/1 forms.) The list_queries/0 function lists query file names.

Examples

iex> PropertyGraph.read_graph("default.cypher")
#GraphCommons.Graph<type: property, file: "default.cypher", data: "CREATE (a)-[:EX]...">

iex> "..." |> PropertyGraph.write_graph("my.test")
#GraphCommons.Graph<type: property, file: "my.test", data: "...">

Types

graph_data()

Specs

graph_data() :: GraphCommons.Graph.graph_data()

graph_file()

Specs

graph_file() :: GraphCommons.Graph.graph_file()

graph_file_test()

Specs

graph_file_test() :: GraphCommons.file_test()

graph_path()

Specs

graph_path() :: GraphCommons.Graph.graph_path()

graph_t()

Specs

graph_t() :: GraphCommons.Graph.t()

graph_type()

Specs

graph_type() :: GraphCommons.Graph.graph_type()

graph_uri()

Specs

graph_uri() :: GraphCommons.Graph.graph_uri()

query_data()

Specs

query_data() :: GraphCommons.Query.query_data()

query_file()

Specs

query_file() :: GraphCommons.Query.query_file()

query_file_test()

Specs

query_file_test() :: GraphCommons.file_test()

query_path()

Specs

query_path() :: GraphCommons.Query.query_path()

query_t()

Specs

query_t() :: GraphCommons.Query.t()

query_type()

Specs

query_type() :: GraphCommons.Query.query_type()

query_uri()

Specs

query_uri() :: GraphCommons.Query.query_uri()

Functions

graph_create(arg)

Delegates to PropertyGraph.Service.graph_create/1

graph_delete()

Delegates to PropertyGraph.Service.graph_delete/0

graph_info()

Delegates to PropertyGraph.Service.graph_info/0

graph_read()

Delegates to PropertyGraph.Service.graph_read/0

graph_update(arg)

Delegates to PropertyGraph.Service.graph_update/1

list_graphs(graph_file_test \\ :exists?)

Specs

list_graphs(graph_file_test()) :: [String.t()]

Lists graphs in the graphs library.

list_graphs_dir(dir, graph_file_test \\ :exists?)

Specs

list_graphs_dir(String.t(), graph_file_test()) :: [String.t()]

Lists graphs in a directory of the graphs library.

list_queries(query_file_test \\ :exists?)

Specs

list_queries(query_file_test()) :: [String.t()]

Lists graphs in the graphs library.

list_queries_dir(dir, query_file_test \\ :exists?)

Specs

list_queries_dir(String.t(), query_file_test()) :: [String.t()]

Lists graphs in a directory of the graphs library.

new_graph(graph_data)

Specs

new_graph(graph_data()) :: graph_t()

Creates a new graph.

new_graph(graph_data, graph_file)

Specs

new_graph(graph_data(), graph_file()) :: graph_t()

new_query(query_data)

Specs

new_query(query_data()) :: query_t()

Creates a new query.

new_query(query_data, query_file)

Specs

new_query(query_data(), query_file()) :: query_t()

query_graph(arg)

Delegates to PropertyGraph.Service.query_graph/1

query_graph(arg, params)

Delegates to PropertyGraph.Service.query_graph/2

query_graph!(arg)

Delegates to PropertyGraph.Service.query_graph!/1

query_graph!(arg, params)

Delegates to PropertyGraph.Service.query_graph!/2

read_graph(graph_file)

Specs

read_graph(graph_file()) :: graph_t()

Reads a graph from the graphs library.

read_query(query_file)

Specs

read_query(query_file()) :: query_t()

Reads a graph from the graphs library.

write_graph(graph_data, graph_file)

Specs

write_graph(graph_data(), graph_file()) :: graph_t()

Writes a graph to the graphs library.

write_query(query_data, query_file)

Specs

write_query(query_data(), query_file()) :: query_t()

Writes a graph to the graphs library.